Moved `_RandomVariableList` into `probnum.prob`
