+++ src/probnum/diffeq/odesolution.py
@@ -7,6 +7,7 @@ Can function values can also be accessed by indexing.
 import numpy as np
 
 from probnum.prob import RandomVariable, Normal
+from probnum.prob.randomvariablelist import _RandomVariableList
 
 
 class ODESolution:
@@ -126,21 +127,3 @@ class ODESolution:
             return f_rvs
         else:
             raise ValueError("Invalid index")
-
-
-class _RandomVariableList(list):
-    @property
-    def mean(self):
-        return np.stack([rv.mean() for rv in self])
-
-    @property
-    def cov(self):
-        return np.stack([rv.cov() for rv in self])
-
-    @property
-    def var(self):
-        return np.stack([rv.distribution.var() for rv in self])
-
-    @property
-    def std(self):
-        return np.stack([rv.distribution.std() for rv in self])
